api-gateway/src/requirements.txt
Flask==2.3.2
requests==2.31.0
PyJWT==2.8.0
